[Ref]: TopMovie handler 코드 분리
